requirements.txt
PyAudio==0.2.9
Wave==0.0.2
argparse==1.2.1
cycler==0.10.0
matplotlib==1.5.1
numpy==1.10.4
pyparsing==2.1.0
pyserial==3.0.1
python-dateutil==2.5.0
pytz==2015.7
requests==2.9.1
six==1.10.0
wsgiref==0.1.2
